Switch charfuncs to returning number of read/written characters
The charfuncs prototypes were returning 0 for success ("I wrote all the
characters you asked me to") or 1 for failure. When we get to verifying
signatures over data we may not know how much to read, so change them to
returning 0 on error/no more data and otherwise the count of the amount
of data read/written.

+++ armor.h
@@ -31,7 +31,7 @@
  * This function ASCII armors a list of OpenPGP packets and outputs it
  * using putchar_func.
  */
-int armor_openpgp_stream(int (*putchar_func)(void *ctx, size_t count,
+int armor_openpgp_stream(size_t (*putchar_func)(void *ctx, size_t count,
 						void *c),
 				void *ctx,
 				struct openpgp_packet_list *packets);
@@ -46,7 +46,7 @@ int armor_openpgp_stream(int (*putchar_func)(void *ctx, size_t count,
  * armored OpenPGP stream and outputs the data as a linked list of
  * packets.
  */
-int dearmor_openpgp_stream(int (*getchar_func)(void *ctx, size_t count,
+int dearmor_openpgp_stream(size_t (*getchar_func)(void *ctx, size_t count,
 					void *c),
 				void *ctx,
 				struct openpgp_packet_list **packets);
