+++ parsekey.h
@@ -59,7 +59,7 @@ int debug_packet(struct openpgp_packet *packet);
  *	then only the public key component of the last key will be returned,
  *	none of the other packets of the key will be read.
  */
-onak_status_t read_openpgp_stream(int (*getchar_func)(void *ctx, size_t count,
+onak_status_t read_openpgp_stream(size_t (*getchar_func)(void *ctx, size_t count,
 				void *c),
 				void *ctx,
 				struct openpgp_packet_list **packets,
@@ -74,7 +74,7 @@ onak_status_t read_openpgp_stream(int (*getchar_func)(void *ctx, size_t count,
  *	This function uses putchar_func to write characters to an OpenPGP
  *	packet stream from a linked list of packets.
  */
-onak_status_t write_openpgp_stream(int (*putchar_func)(void *ctx, size_t count,
+onak_status_t write_openpgp_stream(size_t (*putchar_func)(void *ctx, size_t count,
 						void *c),
 				void *ctx,
 				struct openpgp_packet_list *packets);
